gets notes playing
